requirements.txt
anyio==4.4.0
certifi==2024.6.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.1
pytest-asyncio==0.23.7
python-dotenv==1.0.1
sniffio==1.3.1
tenacity~=8.3.0